Changed subsampling percentages

+++ subsample/scripts/run.sh
@@ -11,25 +11,31 @@ DATA=data/$BASE.fsa
 RES=results/$BASE.subsample.tsv
 REPLICATES=100
 
-gunzip -c -d $DATAGZ > $DATA
+if [ ! -e $DATA ]; then
+    gunzip -c -d $DATAGZ > $DATA
+fi
 
-echo -e "fraction\tprogram\tabundance" > $RES
-for f in 1.0 1.5 2.5 5.0; do
-    for n in $(seq 1 $REPLICATES); do
-        for p in usearch8 vsearch; do
-            
-            $p --fastx_subsample $DATA \
-                --sample_pct $f \
-                --sizein \
-                --sizeout \
-                --threads $THREADS \
-                --fastaout results/temp.fasta
-            t=$(head -1 results/temp.fasta | cut -d= -f2 | cut -d\; -f1)
-            echo -e "$f\t$p\t$t" >> $RES
-            rm results/temp.fasta
-            
+if [ ! -e $RES ]; then
+
+    echo -e "fraction\tprogram\tabundance" > $RES
+
+    for f in 1 2 3 5; do
+        for n in $(seq 1 $REPLICATES); do
+            for p in usearch8 vsearch; do
+
+                $p --fastx_subsample $DATA \
+                    --sample_pct $f \
+                    --sizein \
+                    --sizeout \
+                    --threads $THREADS \
+                    --fastaout results/temp.fasta
+                t=$(head -1 results/temp.fasta | cut -d= -f2 | cut -d\; -f1)
+                echo -e "$f\t$p\t$t" >> $RES
+                rm results/temp.fasta
+            done
         done
     done
-done
 
-cat data/perfect.tsv >> $RES
+fi
+
+#cat data/perfect.tsv >> $RES
